Renamed hoc's
